requirements.txt
Django == 3.2.6
python_dotenv == 0.19.0
requests == 2.22.0
whitenoise==5.3.0
gunicorn==20.1.0
psycopg2==2.9.1